+++ client/src/router/Router.tsx
@@ -8,7 +8,7 @@ import ExamCreate from '../pages/ExamCreate';
 import FindId from '../pages/FindId/FindId';
 import FindPw from '../pages/findPw/FindPw';
 import Home from '../pages/home/Home';
-import Login from '../pages/Login';
+import Login from '../pages/login/Login';
 import MyPage from '../pages/MyPage';
 import PageNotFound from '../pages/PageNotFound';
 import Problem from '../pages/Problem';
